some manifest changes
AndroidManifest.xml changed to support lower API devices
