+++ app/src/main/java/interviewd/Tab2.java
@@ -1,4 +1,4 @@
-package Interviewd;
+package interviewd;
 
 import android.content.Context;
 import android.net.Uri;
@@ -99,7 +99,7 @@ public class Tab2 extends Fragment {
      * <p>
      * See the Android Training lesson <a href=
      * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
+     * >Communicating with Other Fragments</a> for info information.
      */
     public interface OnFragmentInteractionListener {
         // TODO: Update argument type and name
