+++ app/src/main/java/interviewd/Tab4.java
@@ -1,4 +1,4 @@
-package Interviewd;
+package interviewd;
 
 import android.content.Context;
 import android.net.Uri;
@@ -7,22 +7,16 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.animation.Animation;
-import android.view.animation.AnimationUtils;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
-
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link Tab1.OnFragmentInteractionListener} interface
+ * {@link Tab4.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link Tab1#newInstance} factory method to
+ * Use the {@link Tab4#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Tab1 extends Fragment{
+public class Tab4 extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -34,11 +28,7 @@ public class Tab1 extends Fragment{
 
     private OnFragmentInteractionListener mListener;
 
-    Button recordButton;
-    View v;
-
-
-    public Tab1() {
+    public Tab4() {
         // Required empty public constructor
     }
 
@@ -48,11 +38,11 @@ public class Tab1 extends Fragment{
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Tab1.
+     * @return A new instance of fragment Tab4.
      */
     // TODO: Rename and change types and number of parameters
-    public static Tab1 newInstance(String param1, String param2) {
-        Tab1 fragment = new Tab1();
+    public static Tab4 newInstance(String param1, String param2) {
+        Tab4 fragment = new Tab4();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -73,12 +63,7 @@ public class Tab1 extends Fragment{
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-
-        v = inflater.inflate(R.layout.fragment_tab1, container, false);
-
-        setupAll(v);
-
-        return v;
+        return inflater.inflate(R.layout.fragment_tab4, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
@@ -99,15 +84,11 @@ public class Tab1 extends Fragment{
         }
     }
 
-
-
     @Override
     public void onDetach() {
         super.onDetach();
         mListener = null;
     }
-
-
 
     /**
      * This interface must be implemented by activities that contain this
@@ -117,82 +98,10 @@ public class Tab1 extends Fragment{
      * <p>
      * See the Android Training lesson <a href=
      * "http://developer.android.com/training/basics/fragments/communicating.html"
-     * >Communicating with Other Fragments</a> for more information.
+     * >Communicating with Other Fragments</a> for info information.
      */
     public interface OnFragmentInteractionListener {
         // TODO: Update argument type and name
         void onFragmentInteraction(Uri uri);
     }
-
-
-    public void startRecording(){
-
-
-    }
-
-    private void setupAll(View v){
-        setupListeningTextView(v);
-
-
-        updateAll(v);
-    }
-
-
-
-    private void setupListeningTextView(View v){
-        TextView listeningTextV = (TextView) v.findViewById(R.id.listeningTextView);
-        listeningTextV.setVisibility(View.INVISIBLE);
-    }
-
-    //============================================================================================== Update
-
-
-    private void updateAll(View v){
-        updateRecordingButton(v);
-        updateSomething(v);
-
-    }
-
-    private void updateRecordingButton(View v){
-        final Button recordButton = (Button) v.findViewById(R.id.recordButton);
-        final ImageView mainImage= (ImageView) v.findViewById(R.id.centreIcon);
-        recordButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                recordButton.setTextColor(getResources().getColor(R.color.nowRecordingColor));
-                recordButton.setText(getResources().getString(R.string.now_recording_text));
-                showListeningTextView();
-
-
-                Animation animation1 = AnimationUtils.loadAnimation(getContext(), R.anim.on_record_move_main_icon);
-                mainImage.startAnimation(animation1);
-
-                Animation animation2 = AnimationUtils.loadAnimation(getContext(), R.anim.on_record_move_record_text);
-                recordButton.startAnimation(animation2);
-
-                recordButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-
-                    }
-                });
-
-            }
-        });
-    }
-
-
-    private void updateSomething(View v){
-
-    }
-
-    //============================================================================================== Methods
-
-    private void showListeningTextView(){
-        TextView listeningTextV = (TextView) v.findViewById(R.id.listeningTextView);
-        listeningTextV.setVisibility(View.VISIBLE);
-
-    }
-
-
 }
