+++ app/src/main/res/layout/fragment_tab1.xml
@@ -12,8 +12,9 @@
         android:id="@+id/listeningTextView"
         android:layout_width="324dp"
         android:layout_height="430dp"
-        android:layout_marginBottom="8dp"
         android:layout_marginTop="56dp"
+        android:layout_marginBottom="8dp"
+        android:gravity="center"
         android:text="ListeningText"
         android:textColor="@android:color/white"
         app:layout_constraintBottom_toBottomOf="parent"
