+++ app/src/main/res/layout/fragment_tab2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Interviewd.Tab2">
+    tools:context="interviewd.Tab2">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
