+++ app/src/main/res/values/strings.xml
@@ -1,14 +1,43 @@
 <resources>
     <string name="app_name">Interviewd</string>
 
-    <!-- TODO: Remove or change this placeholder text -->
+    <string-array name="tabs_text_array">
+        <item>Home</item>
+        <item>Script</item>
+        <item>Info</item>
+        <item>Settings</item>
+    </string-array>
 
-    <string name="tab1_name">Home</string>
-    <string name="tab2_name">Script</string>
-    <string name="tab3_name">More</string>
-    <string name="tab4_name">Settings</string>
 
-    <string name="app_title_bar_text">Interviewd</string>
+    <!--==============================Tabs====================================-->
+
+    <!--     Home Tab      -->
+
     <string name="start_recording_text">Start recording</string>
     <string name="now_recording_text">Now recording</string>
+
+    <!--     Script Tab      -->
+
+    <!--     Info Tab      -->
+
+    <string-array name="info_title_text_array">
+        <item>How to Use Interviewd</item>
+        <item>About Interviewd</item>
+        <item>About Us</item>
+        <item>[Example Title]</item>
+    </string-array>
+
+    <string-array name="info_body_text_array">
+        <item>You use it like this</item>
+        <item>Interviewd is an app we created. It is also open-source.</item>
+        <item>Johan Yim - Lead Programmer/Developer\nKrystal Ng - Founder/Financial Manager/Interface Designer \n\nGitHub Link: https://github.com/johanyim/Interviewd</item>
+        <item>Example long body text: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed et malesuada lacus. Donec id pharetra augue, nec interdum dolor. Maecenas mattis justo ut ullamcorper lacinia. Quisque ipsum ex, porta quis est quis, dapibus ullamcorper lorem. Duis scelerisque nisl nec metus rhoncus tempus. Ut auctor metus ac tellus molestie, bibendum semper ante mattis. Phasellus at lorem et urna accumsan ultricies. Etiam ac leo sapien. Nulla ornare risus odio, in viverra sem auctor sit amet. Donec non tincidunt arcu.
+
+Nulla nec ipsum vel tortor convallis commodo ut id massa. Nulla facilisi. Nullam eu finibus metus. Praesent quam elit, consequat eget pulvinar semper, condimentum fringilla nulla. Nunc convallis ac sapien sit amet varius. Mauris ac auctor nisl. Integer dapibus nisl id nisl sodales, ac hendrerit dui scelerisque. Maecenas nisi turpis, lacinia non elit at, efficitur finibus purus. Sed sollicitudin malesuada ligula et sodales. Duis faucibus est enim, a viverra ipsum pellentesque eu. Aenean id enim id urna pharetra iaculis. Pellentesque non nulla euismod, viverra orci eget, pretium lectus. Vestibulum vitae sagittis lorem. Vivamus pharetra suscipit nibh, in facilisis augue auctor ut. Etiam dignissim, dolor iaculis sodales porta, purus metus finibus felis, vitae tristique massa arcu auctor est. Cras tempor, felis sed mattis rutrum, nulla ex fringilla orci, vitae mattis quam quam ac felis.</item>
+    </string-array>
+
+
+    <!--     Settings Tab      -->
+
+
 </resources>
